backend/deltahacks2020/requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
djangorestframework==3.7.1